reddit_scraper/requirements.txt
streamlit==1.24.0
pandas==1.5.3
plotly==5.15.0
praw==7.7.0
pytz==2023.3
python-dotenv==1.0.0 